+++ servicio/src/main/java/com/furukawa/reportsystem/servicio/ApplicationConfig.java
@@ -12,7 +12,7 @@ import javax.ws.rs.core.Application;
  *
  * @author Elliot
  */
-@javax.ws.rs.ApplicationPath("webresources")
+@javax.ws.rs.ApplicationPath("api")
 public class ApplicationConfig extends Application {
 
     @Override
@@ -29,7 +29,7 @@ public class ApplicationConfig extends Application {
      * If required, comment out calling this method in getClasses().
      */
     private void addRestResourceClasses(Set<Class<?>> resources) {
-        resources.add(com.furukawa.reportsystem.servicio.Services.class);
+        resources.add(com.furukawa.reportsystem.servicio.LiderServices.class);
     }
     
 }
